+++ packages/editor/src/components/global-styles/block-link.js
@@ -39,7 +39,8 @@ export function GlobalStylesBlockLink( { path, onPathChange } ) {
 		}
 		if (
 			! path ||
-			( path !== '/blocks' && ! path.startsWith( '/blocks/' ) )
+			( path !== '/blocks' && ! path.startsWith( '/blocks/' ) ) ||
+			path.startsWith( '/blocks/elements' )
 		) {
 			return;
 		}